requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0

# Machine Learning and NLP
transformers==4.36.0
torch==2.1.0
scikit-learn==1.3.2
textblob==0.17.1

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Database
sqlalchemy==2.0.23
sqlite3

# Environment and utilities
python-dotenv==1.0.0
schedule==1.2.0
websockets==12.0
redis==5.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
